+++ src/main/resources/db/migration/V4_1__create_siteprojets_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `siteprojets` (
-  `id` int(10) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `nom` VARCHAR(255) COLLATE utf8mb4_unicode_ci NOT NULL,
   `adresse` VARCHAR(255) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `contact` VARCHAR(255) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
   `responsable` VARCHAR(255) COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `localite_id` int(10) NOT NULL,
+  `localite_id` BIGINT NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_siteprojets_localites`
     FOREIGN KEY (`localite_id`) REFERENCES `localites` (`id`)
